// apps/console/src/pages/organizations/people/__generated__/PeopleListView_peoples.graphql.ts
export type PeopleKind = "CONTRACTOR" | "EMPLOYEE" | "SERVICE_ACCOUNT";
import { FragmentRefs } from "relay-runtime";
export type PeopleListView_peoples$data = {
  readonly id: string;
  readonly peoples: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly additionalEmailAddresses: ReadonlyArray<string>;
        readonly createdAt: string;
        readonly fullName: string;
        readonly id: string;
        readonly kind: PeopleKind;
        readonly primaryEmailAddress: string;
        readonly updatedAt: string;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null | undefined;
    };
  };
  readonly " $fragmentType": "PeopleListView_peoples";
};
export type PeopleListView_peoples$key = {
  readonly " $data"?: PeopleListView_peoples$data;
  readonly " $fragmentSpreads": FragmentRefs<"PeopleListView_peoples">;
};